chore: Move graphl to a single place
