+++ packages/core/modules-sdk/src/remote-query/__fixtures__/get-entities-map.ts
@@ -1,6 +1,6 @@
 import { mergeTypeDefs } from "@graphql-tools/merge"
 import { makeExecutableSchema } from "@graphql-tools/schema"
-import { cleanGraphQLSchema } from "../../utils/clean-graphql-schema"
+import { cleanGraphQLSchema } from "@medusajs/utils"
 
 export function getEntitiesMap(loadedSchema): Map<string, any> {
   const defaultMedusaSchema = `
