+++ packages/core/modules-sdk/src/utils/index.ts
@@ -1,4 +1,2 @@
-export * from "./clean-graphql-schema"
-export * from "./graphql-schema-to-fields"
-export * from "./gql-schema-to-types"
-export * from "./gql-get-fields-and-relations"
+export * from "./linking-error"
+export * from "./convert-data-to-link-definition"
