+++ packages/core/utils/package.json
@@ -36,6 +36,8 @@
     "typescript": "^5.6.2"
   },
   "dependencies": {
+    "@graphql-tools/merge": "^9.0.7",
+    "@graphql-tools/schema": "^10.0.6",
     "@medusajs/types": "^1.11.16",
     "@mikro-orm/core": "5.9.7",
     "@mikro-orm/migrations": "5.9.7",
