+++ packages/core/utils/src/bundles.ts
@@ -17,3 +17,4 @@ export * as PromotionUtils from "./promotion"
 export * as SearchUtils from "./search"
 export * as ShippingProfileUtils from "./shipping"
 export * as UserUtils from "./user"
+export * as GraphQLUtils from "./graphql"
