+++ packages/core/utils/src/common/index.ts
@@ -26,7 +26,6 @@ export * from "./get-node-version"
 export * from "./get-selects-and-relations-from-object-array"
 export * from "./get-set-difference"
 export * from "./graceful-shutdown-server"
-export * from "./graphql-relations-entity-map"
 export * from "./group-by"
 export * from "./handle-postgres-database-error"
 export * from "./is-big-number"
