+++ packages/core/utils/src/graphql/__tests__/graphql-ast.ts
@@ -1,4 +1,4 @@
-import GraphQLParser from "../../joiner/graphql-ast"
+import { GraphQLParser } from "../graphql-parser"
 
 describe("RemoteJoiner.parseQuery", () => {
   beforeEach(() => {
