+++ packages/core/utils/src/graphql/clean-graphql.ts
