+++ packages/core/utils/src/graphql/extract-relations-from-graphql.ts
