+++ packages/core/utils/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./graphql"
 export * from "./api-key"
 export * from "./auth"
 export * from "./bundles"
