+++ packages/modules/index/src/utils/gql-to-types.ts
@@ -1,10 +1,9 @@
 import { join } from "path"
 import { CustomDirectives, makeSchemaExecutable } from "./build-config"
 import {
-  gqlSchemaToTypes as ModulesSdkGqlSchemaToTypes,
   MedusaModule,
 } from "@medusajs/modules-sdk"
-import { FileSystem } from "@medusajs/utils"
+import { FileSystem, gqlSchemaToTypes as ModulesSdkGqlSchemaToTypes, } from "@medusajs/utils"
 import * as process from "process"
 
 export async function gqlSchemaToTypes(schema: string) {
@@ -18,6 +17,7 @@ export async function gqlSchemaToTypes(schema: string) {
     schema: executableSchema,
     filename,
     outputDir: dir,
+    joinerConfigs: MedusaModule.getAllJoinerConfigs(),
   })
 
   const fileSystem = new FileSystem(dir)
